Removed test code for Kickle's movement.
Modified TestLevel to include Kickle as a GameObject. 

Added a function called LuaGetKeyEvent that returns an integer related to a key event. This function is used for scripting input handling.

Changed calls to the script so that not all GameObjects have to share the same Lua functions. Now, if a GameObject doesn't have the function called, nothing happens. 

Gave Kickle's script a HandleUserInput function, and Kickle now moves according to this. There are a few odd numbers used that need to be clarified. 

Kickle no longer has collision detection due to the hard coded movement being removed. May handle collision through scripting and figuring out object's location on map, and what objects around it are.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@123 5e24083e-1465-11df-9d2a-2365f6df44e6