Updated Configuration class so that all of its methods and members are static. Load is called in the main function to initialize the static members because App's constructor relies on Configuration attributes.
Gave GameObject a new member string called m_type that is the "type" of object it is (Slime, Kickle, Money Bag, etc.). 

Edited GameObjectMap to load in the type of the object from the XML file. The new setup for instances in the XML file will be to have a type="whatever" along with the x and y attributes. 

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@145 5e24083e-1465-11df-9d2a-2365f6df44e6