GameObject type, width and height have been added to xml GameObject files. Type was moved to the object file rather than the level file.
Kickle's starting point has been moved.

Collision detection has been completely revamped. Each GameObject now holds an sf::Rect for rectangular collision detection. The GameObject calls a collision function in Level which calls a collision function in the object map. Feel free to change the names of the functions, I couldn't think of anything good.

Had to split the LoadFromFile function up into 2 parts. LoadObjectData() and LoadCollisionData(). Kind of bad  naming again. In addition, I made them private. The splitting had to be done because I needed the y position in order to create the collision box in the load function, but the y position couldn't be computed until the animdata from the xml file was retrieved. Thus, LoadObjectData() is called first. Then the positions are set. Then LoadCollisionData() is called.

It's important to note that collision between GameObjects is very different now. When Kickle collides with a slime, they get "stuck" together. This is normal behavior. The goal is to have some action occur when these 2 objects collide.

Lastly, Kickle's collision box is just 48x48 at the moment. Can be changed in the future if we notice issues. 

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@152 5e24083e-1465-11df-9d2a-2365f6df44e6