Moved the DetectObjectCollision from MoveDir to Update() Cleaned up Update some by making a MovementUpdate() private helper methods changed sf::Rect<float>'s into sf::FloatRect Inside DetectCollision instead of manually comparing the float rects changed it to use sf::Rect::Intersects(); Made it call HandleCollision from a lua script upon detecting object collision's It passes HandleCollision the two objects involved in the collision It *should* call HandleCollision in each collision object's scripts.(I haven't tested that yet)
Issues
We need to deal with the animation frame/frameTime.  Currently, when the death animation is set it could be at any point in it's animation depending on the frame(of the last animation) it was on previously.  This results in kickle frequently skipping portions of the death animation.

Also when they collide and kickle dies he will fall down but continue sliding across the ground to his next grid location.  The will probably be as simple as exposing a StopMovement method for lua to call when he dies.


git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@156 5e24083e-1465-11df-9d2a-2365f6df44e6