Made constants for direction in GameObjectUtilities. Made constants for animation type in Kickle.lua
Fiddled around some more with A.I. and tried a variety of approaches. Found out that in real Kickle, the slimes are bugged and can't get to you if you put your pillars in a certain pattern. I want to do better than that, but it isn't easy. For now, if the Slime gets stuck it keeps trying random directions until it can find one it can go in, but this still needs to be improved. 

I believe some element of randomness is key though. Having the slimes relentlessly pursuing Kickle could be tiring for the player.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@186 5e24083e-1465-11df-9d2a-2365f6df44e6