Modified how Slime collision works ( Ben's advice ) by actually reverting them back to their previous position then reversing so that there is no chance another collision will be detected after the first one. I believe this has fixed the problem with slimes getting "stuck" on each other.
However, they still get off of their axes. I believe this has to do with the slime changing from horizontal to vertical or vice versa, and updating movement while it is still in between tiles. The only reason I can think of this occurring is if the AI's logic function gets called without the object being ready yet. 

Added a Lua function to help with this. Eventually I'm going to write a GameObjectUtility that wraps it all up.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@208 5e24083e-1465-11df-9d2a-2365f6df44e6