I implemented the improved input handling. Made a KeyManager class. Each GameObject now has an array of KeyTime->String pairs. Thinking about it now, I probably should have used std: instead of making a the KeyTime. :\ nbd... I guess I just didn't like doing this: std::pair< std::pair<sf::Key::Code,float>, std::string >
Exposed IsMoving to lua as I needed it to get the script working correctly.
In Kickle's script I cleaned up quite a bit, removing pointless variables and changing names.  I renamed mode to state and used it instead of InAction.  Now when you tap a direction Kickle will only face that direction.  If you hold it down for a longer amount of time(0.15 seconds currently) he will walk in that direction.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@210 5e24083e-1465-11df-9d2a-2365f6df44e6