All objects will now have collision rectangles, with grid-based collision as an option (as specified in XML).
Pillars are no longer created if a Slime is in the way. It would be useful to write a function so a GameObject can delete itself in its script. 

XML has been modified for cleaner object files (per advice from Ben). Nathan, I didn't touch your input stuff so just modify them when you get the chance, or if you want me to do it let me know.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@230 5e24083e-1465-11df-9d2a-2365f6df44e6