Fixed some bugs regarding IceBlocks how they would be created when Kickle was facing a solid tile and stuff like that. IceBlock has grid collision now. It shouldn't be having Kickle reverse on it.
For some reason before I decided to commit I noticed that the tiles move all the way when hit, but I had it before so that wasn't happening. Not sure what happened, but I'm about to collapse so can't fix it now.

The Game qualifier is now Level for lua functions in Level.

Ben, note that I added lua functions in Level to GetTile and SetTile. I also made a comment for you in the IceBlock script. You know the tileMap stuff best so I figured you should implement the Iceblock creating an ice tile when it hits water.

LuaAppFuncts removed because they are no longer needed.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@245 5e24083e-1465-11df-9d2a-2365f6df44e6