Replaced BasicAnim_48x60 with KickleAnim_48x60.xml.
Changed the way reversing an animation is handled.  Now when you specify an animation to can specify a <playback></playback> tag that states whether it is played forward or backward.  

With this I added a new animation to the PillarAnim_48x48.xml file that described the backward animation.  

I also added the reverse animations for kickle lowering of the pillars.  

When I put in the playback tag I made it, as well as the looped and individual_times tags, case insensitive.

And you can now opt to leave them out of the <strip> tags altogether.  In which case they are defaulted to forward/true/false correspondingly.

In the future I plan on making many tags option.  Defaulting their values whenever possible.  I've also thought of ways to make some tags unnecessary.

Now time to work on the new improvement to the input that Alex requested.

Oh! Oh!  I fiddled with the pillar lowering frametimes so they match up to kickle's animation.  Now the pillars raise when his hands move up.  This looks a lot better :D

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@247 5e24083e-1465-11df-9d2a-2365f6df44e6