Didn't do much. Just removed a warning. Updated a few things, and put in Maxim's new IceBreath, updating its animation file, etc.
I told Maxim about making sprite sheets and he says he has got it covered in the future. In addition, we were talking about why his work causes the game to crash if we don't re-save it, and he is going to try to figure it out.

IMPORTANT: I've been thinking about some immediate hurdles we have to jump, regarding spawn points and slimes being hit by the ice block. I added them to the TODO, but I'll post about it here...

For spawn points, we should have an animation for the slime coming out, then after it is over spawn the slime on the tile. The big problem is that we need to somehow link slimes to spawn points. If we can do this then checking if the slime exists and spawning a new one if it doesn't in the spawn point's script will be easy.

The second hurdle is slimes being hit by ice blocks. I thought about writing some lua functions to get an object's speed and adjust it. Then change the slime's direction, abandon its normal AILogic, set its animation to freaking out, etc. Another, possibly simpler idea, is that when an ice block hits a slime we could kill the slime on the spot and then have a separate animation in the IceBlock sheet for the slime freaking out along the side of the ice block (see original kickle). More work for the artists, but would give us a lot more flexibility in how we want it to look.

PLEASE post your ideas in this revision. Thanks, and good work everybody. 

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@252 5e24083e-1465-11df-9d2a-2365f6df44e6