Render order actually isn't a problem for now since enemies will throw bags anyway and Kickle will pick them up.
Put the basic mechanics of points in. I actually think we should handle points a bit differently than the original game. The level will still have a timer, that melts (or however the artists do it) as the level continues. The points themselves are held by the WorldState so you can see your total point accumulation in the overworld. How the original game handled points was that it was just the time left on the timer * 10 at the end of the level. Thus, I think we should just render the timer in the actual level, and at the end do a little animation showing the rest of it becoming points in your bank for the overworld. 

We are pretty close to being done with LevelState in general. We need to start seriously designing how the overworld is going to work.

Also I removed the TODO Art list. We should put our art needs on the noobtoob wiki page @ http://wiki.noobtoob.com/Kickle_Cubicle_Level_Editor/Art_To_Do

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@281 5e24083e-1465-11df-9d2a-2365f6df44e6