Modified how tiles work. Tiles now have a type, a name, and an id. These are all returned by GetTile in Lua. Tiles can be set via their name.
IMPORTANT!!! TILEMAP USES std::tr1::tuple! IF IT ISN'T COMPILING, YOU NEED TO GET SERVICE PACK 1 AND FEATURE PACK 1 FOR VISUAL STUDIO!

Cleaned up a lot of scripting stuff.

IceBlock works, but the blocks created aren't aesthetically pleasing. There is no easy way to get the correct tile for the location besides a huge, complex branching of if statements. I tried designing some kind of auto-correction in that tiles would have a link to what tiles they should be adjacent to, and they'd fix themselves when that invariant is broken, but it would just end up that one change would lead to another change so that many SetTile calls would lead to a big chain of events. So I asked Maxim to design some kind of "shimmering bridge" to create when an ice block hits water. This way, we don't have to worry about the aesthetics of ice tile connections.

Put tallchick's stuff in. Had to hardcode a bunch of values into the stagestate's constructor. We might want to handle some kind of layout outside of the application code.

Animation tables are gone for now. Also renamed some stuff.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@286 5e24083e-1465-11df-9d2a-2365f6df44e6