Slimes now respawn at the points from which they originally spawned (like real Kickle).
Fiddled around with speeds of animations and stuff to make the game a bit smoother and quicker.

Made some minor changes to how tilemap works (find() rather than indexing into the map).

Animation is back to how it originally was. We'll stick with it until we come up with a fully designed idea, which we'll discuss at the next meeting.

Removed BasicTypeDefs. Sorry, but I saw it as just a waste of a file. 

I noticed some odd behavior such as when an ice block is created and another slime is right next to it, it will go under it, and a few other things. These will be fixed. I noticed one serious bug, stemming from when gameObjectMap loops through its gameObjects to destroy and destroys them. Sometimes an exception is thrown, and when I examine the call stack, both times it has happened at i = 2, and the values for the gameObject in that spot are all over the place. I have a feeling it doesn't exist or something. Will look into it as soon as possible.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@290 5e24083e-1465-11df-9d2a-2365f6df44e6