Added the NoClip feature. When set, the GameObject won't be restricted by grid objects or non-crossable tiles.
Made some modifications to the API and to the collision map. Level now has GetTileInfo, and GameObject has just GetTile. Hopefully that will be a bit more clear. Also changed the 'SOLID/NOT SOLID' with tiles to 'CROSSABLE/NOT_CROSSABLE' because it seems more clear. Changes in code and API have been made accordingly.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@296 5e24083e-1465-11df-9d2a-2365f6df44e6