Configuration has been trimmed down to what should be there. Things removed either shouldn't have existed (xpad, ypad, if we use them they should be gui) or should have been in the proper place (tile stuff is in tile manager) and were moved accordingly (with xml files modified for information, see test_state.xml). This cleaned up a ton of dependencies on Config. I didn't touch the GUI though, since Ben is working on that.
You may notice the screen size is different. Since there is no xpad or ypad, i adjusted the screen size to fit the tiles exactly (720x720). 

As a kind of odd side note, when I was testing the changes, the game wasn't correctly figuring out the tile GameObjects were on. I took a look at the algorithm in the GetTile functions (which had the "pad" parts removed when I was cutting Config crap out) and noticed that they were wrong in the first place. So that is fixed. My guess would be that the person who originally wrote it didn't realize it was wrong b/c the xpad and ypad stuff they included were just enough to calibrate it correctly.

A few files are just modified b/c vim was showing their carriage returns (^M) at the end of every line (apparently something that happens sometimes when working on files in windows) so I took them out.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@357 5e24083e-1465-11df-9d2a-2365f6df44e6