VERY IMPORTANT: 1) This commit has a bug, but I decided to commit it just so it could go in. Mouse detection in button doesn't work correctly, and if you move the mouse too fast when there is a button on screen (just the title state right now) it lags all input detection (significantly). In the meantime, until this is fixed, I've given the button keyboard input. When you get to the title screen, hit enter to get to the test level. Note however, that if you move the mouse quickly keyboard input is lagged as well.
2) When I was trying to fix the bug the SFML in my workspace got all messed up. PLEASE LET ME KNOW if you see weird lines in the test_level, so I can confirm as to whether the problem is in my workspace or the source code.

Onto the commit...

GUI is in. There is an abstract GUI base class that is able to run lua scripts can have input defined for it - i.e. every gui component can have a lua script and have input defined for it in its xml file. So far I've derived two GUI components, Image and Button. See XML files to see how they are defined, etc. (I will write up documentation...eventually). There is a GUIManager to handle GUI in a state.

I've done a bit of scaffolding to set up a working configuration. Since we don't have a statemanager, I've added a few things to app to allow switching between states. I've also modified GameState so that you don't need to define certain parts in a state file if they aren't needed. For example, the title_state has no Tiles, GameObjects, Collision, or Sound - so only GUI is defined in it. The changes I've made in GameState are temporary as well - I'll be giving the managers default constructors instead.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@390 5e24083e-1465-11df-9d2a-2365f6df44e6