Fairly trivial changes, though make sure to look at Issue 13.
Modified collision rectangles on Slime and Kickle (now more forgiving to the player).

Changed scripts to use 'self' instead of the object's name for the first parameter (more python-esque), and thus the table's name is the object's name. 

Kickle can enter GodMode by hitting 'g'. Note that while NoClip allows you to cross uncrossable tiles and objects, it doesn't stop the HandleCollision function from occuring (so Kickle can still die from running into a slime, etc.)

Added a GameObject called StageManager, that is basically a "nothing" object thats purpose is to run its script. Gave the ability to reset the state via backspace (rather than having to exit the app and re-enter when you die) and moved some stuff that was in the Kickle script out to it (the checking if DreamBags are left). May want to use something other than StageManager in the future, like a GUI element or something, but it suits wel for now. To allow objects like this to be made without forcing the user to specify animations, etc. when they aren't appropriate, all GameObject xml tags are optional. This cut back a lot of the error handling, but I think it is worth it in terms of ease of use.

Registered ResetState() function in GameState (i.e. call Game.ResetState()).

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@419 5e24083e-1465-11df-9d2a-2365f6df44e6