Added timer functions to the GameObject API so each GameObject can handle its own timer. The Move function in the GameObject API now returns a boolean value. It returns true if the GameObject can move in the direction it is facing or is already moving, and false if it can't move in the direction it is facing. This removes the need for lastTileX nonsense.
Brought out more features of existing GameObjects...

IceBlocks are now on a 45 second melting timer, once they are finished they are destroyed and new Slimes spawn (no melting animation at the moment, so they just disappear). Also made IceBlock's collision rect a bit smaller.

Made changes to Slime and Penguin's AI so that they don't constantly do the same thing if Kickle isn't moving and they are in a position of fairly limited movement. The AI can still be improved, but it is a work in progress. Since they use the same AI, I moved it out to GameObjectUtilities to a function called GenericEnemyAI.

If an IceBreath hits an IceBlock, it resets its melting timer to zero.

Changed UpdateMovement() in GameObject to UpdateAI. I'm going to change the lua function that is called to just be AI rather than AILogic.



git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@425 5e24083e-1465-11df-9d2a-2365f6df44e6