Finished changes that needed to be made to existing GameObjects...
IceBlocks can now be destroyed when kicking them and they can't move anywhere.

IceBreath freezes Penguins for 5 seconds. If Kickle kicks the Penguin when frozen, it is destroyed.

Added function to GameObjectAPI called BlockTile that takes a bool stating whether the object should block the tile it is on or not.

Fixed a problem with collision detections. The GameObjectManager was looping through all of the objects in the GameObjectManager to see if the object collided with any of them. The problem with that was, if the GameObject was colliding with 2 objects at the same time, only the one first in the list would be detected. Instead of returning a variable number of arguments to Lua (I think returning just one adheres more to the "just works" philosophy. Doesn't make sense how it works, just does), when an object in the list is collided with it is spliced to the end of the list. This is constant time. However, this required a change in how the list is looped through when GameObjectManager is updating collisions. If you use an iterator as your counter, it will be an infinite loop since 2 GameObjects colliding will continually be put at the end of the list. So instead I use the list's size as the counter (but obviously still using iterators to access). This is fine, since it is guaranteed that no GameObjects will be removed from the list until the end of GameObjectManager's Update() function. 

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@427 5e24083e-1465-11df-9d2a-2365f6df44e6