Modified where lua functions are located. I wanted the lua functions to have private access to the manager that they were appropriate with, rather than having to write a getter for everything we wanted to do and having all the funcs in GameState. So I split them out into the separate managers.
I also moved the lua state from GameState to the App, and made it so the state is constant throughout, rather than creating a new one after every GameState (don't worry, all the gunk under the hood gets cleaned up after every state change). In turn, I gave GameState and the managers a RegisterLuaFuncs method so they don't register their functions for a single state multiple times.

Added a lua function called GetGameObjects that returns all objects in the manager of a given type in a table. This is going to be modified in the future to allow one to specify objects in a certain area.

PowerRock behavior is in. Turns all slimes on screen into ice blocks.

GardenLand_D is in.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@430 5e24083e-1465-11df-9d2a-2365f6df44e6