Sped up Kickle, Ice Breaths, and the pillar animation.
About 2 hours into working on the GUI thing, I realized that I had been talking to someone while I was working on it and the design wasn't optimal and would probably be changed, so I reverted all of it. Lesson learned. Title screen doesn't look pretty, but I'm not going to do haphazard stuff for the sake of a tester liking the title screen at this point.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@440 5e24083e-1465-11df-9d2a-2365f6df44e6