Modified Doxyfile to allow JavaDoc commenting.
Created a more encapsulated relationship between ObjectManager and Object using the Attorney-Client idiom. The only class in the entire application that actually touches Objects is ObjectManager, so it didn't make much sense to have public member functions in Object when only ObjectManager was using it. So I decided to make most (Lunar needs a few things, but this will be changed soon) of Object's stuff private and ObjectManager a friend. However, I didn't like that it now had access to EVERYTHING in Object (especially the helper functions). I wanted more granularity, not the all or nothing that friend gives. So I used the Attorney-Client idiom to establish a more distinct contract between ObjectManager and Object. See the code for details (and the link provided), it is fairly self explanatory. Now, Object is totally encapsulated except for a few things only available to ObjectManager (oh and of course ObjectAttorney :P). My only gripe is that using ObjectAttorney::Whatever in GameObjectManager for everything is a bit verbose, but it is more than worth it.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@464 5e24083e-1465-11df-9d2a-2365f6df44e6