Fixed NoClip bug where Object could move off of map.
Before that was fiddling with the sprite sheet clip size problem, so there are a few unrelated changes, but they don't break anything. The spring animation file is just "correct" now and I put a spring in GardenLand_E (in its appropriate spot, GardenLand_E has springs). You can see that the spring is being animated one tile above where it should be.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@465 5e24083e-1465-11df-9d2a-2365f6df44e6