Fixed bug where objects whose sheet clips were smaller than the tile size were being blitted to the wrong tile. See Issue 11 for more details.
Renamed Spring stuff appropriately, and added them to GardenLand_E (but no behavior right now).

Objects now hold their tile location instead of recalculating it on every call of GetTileX and GetTileY. Tile location changes when they have traveled halfway across the current tile. 

Fixed a potential bug in which there was an initialization like this in ObjectManager's LoadData -- int x, y, strip = 0; This only assigns strip to 0, and leaves x and y uninitialized (won't necessarily initialize to 0). This revealed itself when implementing Objects holding their own tile locations, but it is fixed now, and I scanned through the rest of the code to see if there was anything similar.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@466 5e24083e-1465-11df-9d2a-2365f6df44e6