Sorry this took so long, but holy crap! I didn't set this up as much up front as I should have, so ended up changing a lot of things midway, going back on a lot of stuff, etc. I ended up taking a couple day break b/c of an annoying as hell bug that just ended up being an iterator invalidation because of AdjustGridCoord.
Basically, this commit implements the idea of an object grid and map of object types for much faster access in various API function calls. It requires an extra bit of keeping track because objects need to be moved in the object grid, but it's spread out and minimal so shouldn't be a problem (will see if it is via profiling).

P.S. While I was trying to find that bug I valgrinded the program and no leaks! That is pretty damn surprising considering we do absolutely no unit testing.

git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@506 5e24083e-1465-11df-9d2a-2365f6df44e6