Object movement is now frame independent, so their speed is now represented in pixels per second.
AI for enemies is exactly the same as the real Kickle Cubicle.

Added a few float comparison functions to Utilities, some private methods to Object to retrieve opposite direction and velocity vector, removed GetCamera from App, etc.

Added function called OnCollisionCourse that can determine if, given 2 objects' current speed and direction, whether they will collide, or are already colliding. I ended up not using it for what I initially wrote it for, but I plan on replacing our entire collision detection with it in the next few updates. The problem with just determining collision based on two rectangles intersecting is that if one or both of the objects is moving fast enough the collision can be missed completely. This actually happens sometimes now since ObjectManager is a lot more efficient. The OnCollisionCourse method actually obtains how long it will take for the object to reach the other object given both of their current states. Going to read up on it a bit more before I start changing things, but that is the next thing I'm going to be working on.


git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@514 5e24083e-1465-11df-9d2a-2365f6df44e6